PDKIND: Move implementation details from header to implementation file
